Use Long types in client (#94)

+++ android/app/src/main/java/ru/tesseract/assets/domain/GeneralAssetInfo.kt
@@ -12,9 +12,9 @@ class GeneralAssetInfo(
     @SerialName("companyTitle")
     val companyTitle: String,
     @SerialName("assetPrice")
-    val price: Int,
+    val price: Long,
     @SerialName("assetPriceDiff")
-    val priceDiff: Int,
+    val priceDiff: Long,
     @SerialName("favouriteStatus")
     val isFavorite: Boolean,
 )
