+++ android/app/src/main/java/ru/tesseract/assets/ui/AssetPrice.kt
@@ -1,3 +1,3 @@
 package ru.tesseract.assets.ui
 
-fun formatPrice(value: Int) = "%.2f ₽".format(value.toDouble() / 100)
+fun formatPrice(value: Long) = "%.2f ₽".format(value.toDouble() / 100)
